+++ skypro-react/src/homework_6/Input.js
@@ -1,29 +1,7 @@
-import { useState } from 'react';
-
-const useInputReq = (initialValue) => {
-  const [inputValue, setInputValue] = useState(initialValue);
-  const [isError, setIsError] = useState(false);
-
-  const onBlurChange = () => {
-    setIsError(!inputValue && true);
-  };
-
-  const onValueChange = (event) => {
-    const targetVal = event.target.value;
-    setInputValue(targetVal);
-    setIsError(inputValue && false);
-  };
-
-  return {
-    inputValue,
-    isError,
-    onBlurChange,
-    onValueChange,
-  };
-};
+import useInputRequired from './useInputRequired';
 
 const Input = ({ type, name, text }) => {
-  const inputCopy = useInputReq('');
+  const inputCopy = useInputRequired('');
 
   return (
     <>
